% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_data_with_zero.R
\name{bin_data_with_zero}
\alias{bin_data_with_zero}
\title{Create Dataframe for Binned Data with a Dedicated Zero Bin}
\usage{
bin_data_with_zero(
  df,
  value_col = "value",
  group_col = NULL,
  bins = 50,
  cutoff_low = NULL,
  cutoff_high = NULL
)
}
\arguments{
\item{df}{A data frame containing the data to be binned.}

\item{value_col}{The name of the column with numeric values to bin.}

\item{group_col}{The name of the columns to group by (e.g., for comparison). This is defaulted to NULL}

\item{bins}{Number of bins to create (default is 50).}

\item{cutoff_low}{Optional lower number cutoff (numeric). Default is 5th percentile}

\item{cutoff_high}{Optional upper number cutoff (numeric). Defailt is 95th percentile}
}
\value{
A binned data frame with columns for the bin, counts, proportions, and each grouping column
}
\description{
This function creates bins for continuous values to visualize into a histogram-like format
with a dedicated bin for zero. It is an input into visualization functions to display
the shape of a continuous variable with a dedicated option for values that are 0.
}
\examples{
binned <- bin_data_with_zero(data_frame, value_col = "price", group_col = "group")
}
