% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize_zero_bin_legend.R
\name{customize_zero_bin_legend}
\alias{customize_zero_bin_legend}
\title{Format Legend Names}
\usage{
customize_zero_bin_legend(
  p,
  binned_data,
  group_col = "name",
  legend_labels = NULL,
  fill_colors = NULL
)
}
\arguments{
\item{p}{The histogram created by \code{plot_zero_bin_histogram}}

\item{binned_data}{The data used to create p, made by \code{bin_data_with_zero}.}

\item{group_col}{A character vector of the grouping variables for the legend.}

\item{legend_labels}{A character vector of new names for the legend. # Named vector: c("old_name1" = "New Label 1", ...)}

\item{fill_colors}{A vector of colors to map to each group. # Named vector: c("New Label 1" = "#color1", ...)}
}
\value{
A data frame with an updated \code{name_legend} column and color mapping.
}
\description{
Applies custom names and color values to a column for use in legend formatting.
}
\examples{
df <- format_legend_names(df, c("Real", "Model"), c("Real World", "Model"), c("blue", "orange"))
}
