% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zero_bin_histogram.R
\name{plot_zero_bin_histogram}
\alias{plot_zero_bin_histogram}
\title{Plot Zero Bin Histogram}
\usage{
plot_zero_bin_histogram(
  binned_data,
  group_col = "name",
  counter = 0.5,
  y_var = "prop",
  y_format = percent_format(),
  x_format = number_format(),
  drop_line = F
)
}
\arguments{
\item{binned_data}{A data frame created by \code{bin_data_with_zero}.}

\item{group_col}{Column name for grouping (fill aesthetic).}

\item{counter}{Numeric value to control x-axis label spacing.}

\item{y_var}{Column to use on the y-axis, such as "prop" or "obs_num". Default is "prop" which gives the proportion}

\item{y_format}{Format for all numeric values on the y-axis (function). Default is percent_format()}

\item{x_format}{Format for all numeric values on the x-axis (function). Default is numeric_format()}

\item{drop_line}{Flag to create a dashed drop line at 0 on the x-axis to indicate "No change"}
}
\value{
A \code{ggplot} object representing the histogram.
}
\description{
Generates a ggplot histogram with a zero-centered bin and adjusted axis labels.
}
\examples{
p <- plot_zero_bin_histogram(binned_data)
print(p)
}
